Rename commonjs entry point file
