+++ source/likely.js
@@ -2,7 +2,7 @@
 // It’s written with CommonJS imports and exports to make possible doing `module.exports = likely`.
 // This is required so that users work with `window.likely`, not `window.likely.default`
 
-const likely = require('./index.js');
+const likely = require('./likely-commonjs.js');
 
 window.addEventListener('load', () => {
     likely.initiate();
